little modifications

+++ deeplib/initializers.py
@@ -6,20 +6,20 @@
 #    By: thflahau <thflahau@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2020/12/06 15:03:31 by thflahau          #+#    #+#              #
-#    Updated: 2020/12/24 13:22:17 by thflahau         ###   ########.fr        #
+#    Updated: 2021/01/11 20:09:37 by thflahau         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 import numpy.random as nrand
 
-def regular(shape):
-	return nrand.ranf(size=shape).astype(float)
+def regular(shape, dtype=float):
+	return nrand.ranf(size=shape).astype(dtype)
 
-def regular_scaled(shape):
-	return nrand.ranf(size=shape).astype(float) * 0.01
+def regular_scaled(shape, dtype=float):
+	return nrand.ranf(size=shape).astype(dtype) * 0.01
 
-def uniform(shape):
-	return nrand.uniform(-1.0, 1.0, size=shape).astype(float)
+def uniform(shape, dtype=float):
+	return nrand.uniform(-1.0, 1.0, size=shape).astype(dtype)
 
-def normal(shape):
-	return nrand.normal(size=shape).astype(float)
+def normal(shape, dtype=float):
+	return nrand.normal(size=shape).astype(dtype)
